first steps...
This is just about enough wrapping to produce some messages to a topic.
I'm struggling with an odd sort of bug, which seems to unpredictably
make the Topic.produce() call error out.  Starting with something like

    import librdkafka
    c = librdkafka.Config()
    c.set("metadata.broker.list", "kafka0:9092")
    p = librdkafka.Producer(c)
    tc = librdkafka.TopicConfig()
    t = librdkafka.Topic("newtopic", p, tc)

I can repeatedly safely do

    t.produce("whatever")

but curiously neither of

    for _ in xrange(1): t.produce("whatever")
    if True: t.produce("whatever")

both of which raise

    LibrdkafkaException: Broker: Message size too large

(note: it's the same message?!).  Even more curiously, the following things do
work just fine:

    x = lambda: t.produce("whatever")
    for _ in xrange(1): x()
    for _ in xrange(100000): x()
    for _ in xrange(100):
        for __ in xrange(100):
            t.produce("whatever")

Will have to dig into this.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>