improved msg_opaque handling
This does away with the uuid solution of the parent commit, resulting in
a threefold speedup (I tested this only in passing, by setting the
number of messages, n, to a high number in MsgOpaquesTestCase).

In addition, the test case now also (implicitly) requires that
non-hashable objects such as lists may be passed as msg_opaques (this
worked fine in the parent commit, but wasn't tested for, and almost
didn't work in the new code).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>