fig: new kafka/zk images, kafka 0.8.2
This comes with a number of improvements.  The kafka container is now
available from Docker Hub as an automated build, and it's been updated
to kafka 0.8.2, but more importantly, it now allows one to use fig-scale
to test with multiple brokers.  The naming-hacks in fig.yml were no
longer needed and have been dropped.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>