partition_reader: overhaul of toppar locking
The functional change here is mostly that locked toppars are no longer
held in a global set, but stored on their parent Consumer-handle
(through a new TopparManager).  Storing them on a per-Consumer basis,
and in addition, tracking the topics by name, makes it much easier to
reason about correctness.

The other aim here though, was to make the object referencing simpler so
that the garbage collector might actually be able to clean up
KafkaHandles correctly before exit (see #6).  This turns out not to be a
sufficient improvement for that.

Still, closes #10.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>