partition_reader: restructure to use finaliser
This is the second overhaul of the same module in a few commits' time;
the first time around I broke up the circular references that got in the
way of unwinding object state, but it turns out I got it wrong and broke
the wrong half of the circle: in logger output, it would now show that
the queues would at times outlive the topic- and kafka-handles.

Here, we re-introduce references to topic, so as to ensure the topic
handles outlive the queues and partitions.  We also define
BaseTopic.__hash__ in such a way that toppars opened for reading shall
really be unique.

Having done all this, we still see some problems winding down state -
the difference between the tests that work and the tests that don't [*]
is that the latter have some state set up in setUpClass() - ie handles
that are stored on the class, not the instance.  So... to be continued!

[*] In fact, all tests pass, but in the context of #6, when I say
'work', I mean that rd_kafka_wait_destroyed() gets called successfully -
in the worst kind of failure, it doesn't get called at all - and that it
doesn't time-out.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>