tests: fix hack where we used invalid partition_ids
This turns out to confuse the queue (or the topic) enough that it would
be impossible to shut it down.  The test we fix here *still* doesn't
unwind cleanly: only when we comment-out the consume() calls at the end
does the patch we commit here make any difference - ie there's more to
do here.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>