Producer: add timeout to finaliser
Sorry, I'm splitting my commits the wrong way around: this should go
*before* 00c75379f121444e1f07988a628e581d93c7c543 (which will now hang
indefinitely after completing tests!).

If the producer queue is stuck (in this case, due to issue #13), so that
the queue never gets to zero, we'd loop indefinitely in the finaliser.
The only reason that that wasn't apparent before 00c753, was that the
finaliser wouldn't be called at all!

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>